refactor: 공통 상수와 유틸을 lib으로 이동(#95)
- APPLICATION_STATUS_META/ORDER를 lib/constants/application-status로 분리
- PLATFORM_LABEL을 lib/constants/job-platform으로 통합
- formatAppliedAt을 lib/utils/date로 이동 후 중복 파일 제거
