+++ lib/constants/job-platform.ts
@@ -1,0 +1,8 @@
+import type { JobPlatform } from "@/lib/types/job";
+
+export const PLATFORM_LABEL: Record<JobPlatform, string> = {
+  LINKEDIN: "LinkedIn",
+  MANUAL: "직접 입력",
+  SARAMIN: "사람인",
+  WANTED: "원티드",
+};
